IDEA-71725 Replace structurally -> Replace all does nothing

+++ plugins/structuralsearch/source/com/intellij/structuralsearch/JavaReplaceHandler.java
@@ -386,10 +386,8 @@ public class JavaReplaceHandler extends StructuralReplaceHandler {
     else {
       final PsiElement nextSibling = el.getNextSibling();
       el.delete();
-      if (nextSibling.isValid()) {
-        if (nextSibling instanceof PsiWhiteSpace) {
-          nextSibling.delete();
-        }
+      if (nextSibling instanceof PsiWhiteSpace && nextSibling.isValid()) {
+        nextSibling.delete();
       }
     }
